% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{beta.degree}
\alias{beta.degree}
\title{Convert beta (slope of regression line) to degrees}
\usage{
beta.degree(beta)
}
\arguments{
\item{beta}{slope of regression line}
}
\value{
angle
}
\description{
Convert beta (slope of regression line) to degrees
}
\examples{
\dontrun{ 
beta.degree(1)
}
}
\references{
\url{http://r.789695.n4.nabble.com/slope-calculation-td858652.html }
}
