% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports 2.R, R/RcppExports.R
\name{blocksum}
\alias{blocksum}
\title{Blocksum}
\usage{
blocksum(r, c)

blocksum(r, c)
}
\arguments{
\item{r}{An integer vector}

\item{c}{The number of observations in each block}

\item{x}{An integer vector}
}
\value{
The function returns the block sum of the vector.
}
\description{
Blocksum

Compute the block-wise sum of a vector.
}
