% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.r
\name{bt.append.today}
\alias{bt.append.today}
\title{Append today's quotes}
\usage{
bt.append.today(b, data.today)
}
\description{
data.today = getQuote.yahoo.today(ls(data))
	print(data.today)
bt.append.today(data, data.today)
}
