% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.make.trade.event.summary.table}
\alias{bt.make.trade.event.summary.table}
\title{Summarize bt.run.share.ex events}
\usage{
bt.make.trade.event.summary.table(bt, to.text = F)
}
\description{
Summarize bt.run.share.ex events
}
