% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.r
\name{bt.prep.trim}
\alias{bt.prep.trim}
\title{Trim data environment}
\usage{
bt.prep.trim(b, dates = NULL)
}
\arguments{
\item{b}{original environment with symbols time series}

\item{dates}{dates to keep from original environment}
}
\value{
updated environment with symbols time series
}
\description{
This function will remove weights that are smaller than given threshold
}
\examples{
\dontrun{ 
bt.prep.trim(data, endpoints(data$prices, 'months'))
bt.prep.trim(data, '2006::')
}
}
