% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.run.share.ex.invest}
\alias{bt.run.share.ex.invest}
\title{minimally update portfolio to allocate cashflow
only add new positions, such that abs(shares) only increases}
\usage{
bt.run.share.ex.invest(weight.new, weight.prev, weight.change.index, price,
  share, cash, cashflow, commission, lot.size, silent = T,
  control = default.round.lot.control())
}
\description{
minimally update portfolio to allocate cashflow
only add new positions, such that abs(shares) only increases
}
