% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.run.share.ex}
\alias{bt.run.share.ex}
\title{New Share Back Test functionality to properly capture portfolio changes.
This is an event driven back test that updates portfolio at every trade event}
\usage{
bt.run.share.ex(
  b,
  prices = b$prices,
  clean.signal = T,
  trade.summary = F,
  do.lag = 1,
  silent = F,
  capital = 1e+05,
  commission = 0,
  weight = b$weight,
  dates = 1:nrow(b$prices),
  lot.size = 0.01,
  control = list(round.lot = default.round.lot.control()),
  cashflow.control = list(),
  adjusted = T,
  dividend.control = list(foreign.withholding.tax = NULL, invest =
    spl("cash,rebalance,update")),
  tax.control = NULL
)
}
\description{
Summary:
* store actual positions and cash at the end of the day,
hence if we are rebalancing on T, the positions will be shown on T
* for weights, use weights from T-1, to properly reflect weights throughout day
}
\details{
If UnAdjusted logic is enabled, each price history is expected to have Dividend and
Split columns, it can be populated with bt.unadjusted.add.div.split function

events are trades and div/splits for existing positions
* if div => adjust cash for 
 long = cash + share * div * (1-div.tax)
 short = cash - share * div
if split => adjust share by split ratio
}
