% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.run.share.ex.test.cashflow}
\alias{bt.run.share.ex.test.cashflow}
\title{Test for bt.run.share.ex cashflow functionality}
\usage{
bt.run.share.ex.test.cashflow()
}
\description{
Test for bt.run.share.ex cashflow functionality
}
