% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.run.share.ex.test.tax}
\alias{bt.run.share.ex.test.tax}
\title{Test for bt.run.share.ex tax functionality}
\usage{
bt.run.share.ex.test.tax()
}
\description{
Test for bt.run.share.ex tax functionality
}
