% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bt.run.share}
\alias{bt.run.share}
\title{Helper function to backtest for type='share'}
\usage{
bt.run.share(b, prices = b$prices, clean.signal = T, trade.summary = F,
  do.lag = 1, do.CarryLastObservationForwardIfNA = TRUE, silent = F,
  capital = 1e+05, commission = 0, weight = b$weight,
  dates = 1:nrow(b$prices))
}
\description{
Helper function to backtest for type='share'
}

