% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.run.share.unadjusted.test.data}
\alias{bt.run.share.unadjusted.test.data}
\title{Test for bt.run.share.unadjusted functionality}
\usage{
bt.run.share.unadjusted.test.data()
}
\description{
Test for bt.run.share.unadjusted functionality
}
