% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bt.run.weight.fast}
\alias{bt.run.weight.fast}
\title{bt.run - really fast with no bells or whisles
working directly with xts is alot slower, so use coredata}
\usage{
bt.run.weight.fast(b, do.lag = 1, do.CarryLastObservationForwardIfNA = TRUE)
}
\description{
bt.run - really fast with no bells or whisles
working directly with xts is alot slower, so use coredata
}

