% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{bt.unadjusted.add.div.split}
\alias{bt.unadjusted.add.div.split}
\title{Append dividend and split columns}
\usage{
bt.unadjusted.add.div.split(data.raw, yahoo.round.up.nearest.cents = F,
  infer.div.split.from.adjusted = F)
}
\description{
Append dividend and split columns
}
