\name{col.add.alpha}
\alias{col.add.alpha}
\title{Make color semi-transparent}
\usage{
  col.add.alpha(col, alpha = 150)
}
\arguments{
  \item{col}{color(s)}

  \item{alpha}{alpha, \strong{defaults to 150}, for more
  info please see \code{\link{rgb}}}
}
\value{
  new color
}
\description{
  Make color semi-transparent
}
\examples{
\dontrun{
col.add.alpha('gray')
}
}

