% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.calmar}
\alias{compute.calmar}
\title{http://alumnus.caltech.edu/~amir/mdd-risk.pdf
The Calmar Ratio is equal to the compounded annual growth rate divided by the maximum drawdown.
The maximum drawdown is typically measured over a three year period.
Calmar Ratio = CAGR / MAXDD}
\usage{
compute.calmar(x)
}
\description{
http://alumnus.caltech.edu/~amir/mdd-risk.pdf
The Calmar Ratio is equal to the compounded annual growth rate divided by the maximum drawdown.
The maximum drawdown is typically measured over a three year period.
Calmar Ratio = CAGR / MAXDD
}

