% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.DVR}
\alias{compute.DVR}
\title{http://cssanalytics.wordpress.com/2009/10/15/ft-portfolio-with-dynamic-hedging/
DVR is the Sharpe Ratio times the R-squared of the equity curve}
\usage{
compute.DVR(bt)
}
\description{
http://cssanalytics.wordpress.com/2009/10/15/ft-portfolio-with-dynamic-hedging/
DVR is the Sharpe Ratio times the R-squared of the equity curve
}

