% Generated by roxygen2 (4.0.1): do not edit by hand
\name{compute.turnover}
\alias{compute.turnover}
\title{Portfolio turnover
http://wiki.fool.com/Portfolio_turnover
sales or purchases and dividing it by the average monthly value of the fund's assets}
\usage{
compute.turnover(bt, b)
}
\description{
Portfolio turnover
http://wiki.fool.com/Portfolio_turnover
sales or purchases and dividing it by the average monthly value of the fund's assets
}

