\name{create.monthly.table}
\alias{create.monthly.table}
\title{Create monthly table}
\usage{
  create.monthly.table(monthly.data)
}
\arguments{
  \item{monthly.data}{monthly time series}
}
\value{
  matrix with Months as columns and Years as rows
}
\description{
  Transform given monthly time series into matrix with
  Months as columns and Years as rows
}
\examples{
\dontrun{
create.monthly.table(monthly.ret)
}
}

