% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cross}
\alias{cross}
\title{Cross - gives a "1" or true on the day that ARRAY1 crosses above ARRAY2. Otherwise the result is "0".
To find out when ARRAY1 crosses below ARRAY2, use the formula cross(ARRAY2, ARRAY1)
http://www.amibroker.com/guide/afl/afl_view.php?id=34}
\usage{
cross(array1, array2)
}
\description{
Cross - gives a "1" or true on the day that ARRAY1 crosses above ARRAY2. Otherwise the result is "0".
To find out when ARRAY1 crosses below ARRAY2, use the formula cross(ARRAY2, ARRAY1)
http://www.amibroker.com/guide/afl/afl_view.php?id=34
}

