% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{data.ft.search.ticker}
\alias{data.ft.search.ticker}
\alias{data.ft.index.members}
\alias{data.ishares.universe}
\title{[sp500](http://markets.ft.com/research/Markets/Tearsheets/Constituents?s=INX:IOM)
[tsx](http://markets.ft.com/research/Markets/Tearsheets/Constituents?s=TSEA:TOR)
[ftse](http://markets.ft.com/research/Markets/Tearsheets/Constituents?s=FTSE:FSI)}
\usage{
data.ft.search.ticker(search.field = "tsx", sec.type = "IN")

data.ft.index.members(ft.symbol = "INX:IOM", force.download = FALSE,
  data.filename = paste0(gsub(":", "_", ft.symbol), ".Rdata"),
  data.keep.days = 30)


  data.ishares.universe(portfolios = "239726-239623-239458-239566-239706-239763-239708-239637-239710-239467-239774-239565-239665-239826-239707-239500-239725-239695-239718-239644-244049-239451-244050-239452-239728-239456-239454-239465-239561-239719-239626-239766-239699-239572-239463-239717-239714-239855-239712-239709-239455-239600-239627-239563-239762-239650-239764-239723-239536-239520-239482-239724-239466-259622-239775-239681-239659-239641-239612-239534-239773-239605-239615-239499-239628-239736-256101-239686-239522-239622-239601-239854-239464-239468-239674-268708-244048-239690-239619-239594-239511-239657-239607-239737-239744-239670-239512-239507-251614-239508-239685-239741-239524-239768-239772-239506-264617-239516-239423-239683-239513-239505-239746-258100-239713-239757-239460-239769-239453-239580-239664-239543-239514-239761-239715-239750-239510-239830-239756-251616-239758-239716-239540-272532-239502-239771-239450-264619-239740-239523-239457-259623-239519-239579-239720-259624-239678-244051-239501-239509-239731-239582-239503-239661-239667-264615-239765-239545-239680-239649-264623-239517-239751-239521-239729-239689-239705-239648-239588-239618-239528-239462-239688-239692-239669-239684-239730-239677-239581-270319-239675-239668-239739-239733-239696-239459-239518-239645-271054-239585-239767-239742-239671-239606-251465-239583-268704-239676-239586-239584-239753-239748-239430-251474-239745-239752-279626-239666-239550-239614-239424-239662-239722-239655-239734-268752-239526-258098-239610-271056-239831-239429-272342-239654-272341-239663-239629-254263-264507-239759-239587-239829-239461-239621-239551-258510-254551-239504-239672-239721-239642-239515-254553-272343-269394-272824-239609-239738-239529-239539-239544-239431-239527-254555-272340-260973-239660-264503-239537-271544-251476-239770-272822-239693-239443-272344-239735-272346-264273-264613-239656-273753-251477-272345-275382-239445-239653-276546-264542-264275-272112-264544-239613-264611-273746-276544-239570-239652-239651-239530-239525-239647-239620-260652-260975-239673-272819-254562-271540-258806-239552-283378-239691-271538-264127-239444-273743-273775-239638-275389-280052-273771-272825-275397-264606-253433-280049-272823-275384-271542-272821-273750-272820-273763-279286-270316-280048-280051-280050-275399-280769-273748-280771-273766-280774-273759-273768",
  force.download = FALSE, data.filename = "ishares.universe.Rdata",
  data.folder = paste(getwd(), "data.ishares", sep = "/"),
  data.keep.days = 30)
}
\description{
[get industry/sector information from yahoo finance](https://ca.finance.yahoo.com/q/pr?s=RY.TO)
}
\examples{
\dontrun{ 
data.ft.search.ticker('s&p 500')
}
\dontrun{ 
data.ft.index.members('TSEA:TOR')
}
\dontrun{ 
data.ishares.universe()
}
}
