\name{date.dayofweek}
\alias{date.day}
\alias{date.dayofweek}
\alias{date.month}
\alias{date.quarter}
\alias{date.week}
\alias{date.year}
\title{Dates Functions}
\usage{
  date.dayofweek(dates)

  date.day(dates)

  date.week(dates)

  date.month(dates)

  date.quarter(dates)

  date.year(dates)
}
\arguments{
  \item{dates}{collection of dates}
}
\value{
  transformed dates
}
\description{
  Dates Functions
}
\examples{
\dontrun{
date.dayofweek(Sys.Date())
}
}

