% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DV}
\alias{DV}
\title{DV2 indicator (DVB)
http://blog.fosstrading.com/2009/07/david-varadis-rsi2-alternative.html
http://davesbrain.blogs.com/mindmoneymarkets/2010/08/dvib-combo-.html}
\usage{
DV(HLC, n = 2, bounded = FALSE)
}
\description{
DV2 indicator (DVB)
http://blog.fosstrading.com/2009/07/david-varadis-rsi2-alternative.html
http://davesbrain.blogs.com/mindmoneymarkets/2010/08/dvib-combo-.html
}

