% Generated by roxygen2 (4.0.1): do not edit by hand
\name{DVI}
\alias{DVI}
\title{DVI indicator
http://cssanalytics.wordpress.com/2010/07/29/dvi-and-spy-performance/
http://marketsci.wordpress.com/2010/07/29/exploring-the-dvi-indicator-extreme-readings/
http://marketsci.wordpress.com/2010/07/27/css-analytics%E2%80%99-dvi-indicator-revealed/
http://dvindicators.cssanalytics.com/community/?vasthtmlaction=viewtopic&t=47.0
http://quantingdutchman.wordpress.com/2010/07/28/dvi-indicator-for-amibroker/
http://davesbrain.blogs.com/mindmoneymarkets/2010/08/dvib-combo-.html}
\usage{
DVI(x, n = 250)
}
\description{
DVI indicator
http://cssanalytics.wordpress.com/2010/07/29/dvi-and-spy-performance/
http://marketsci.wordpress.com/2010/07/29/exploring-the-dvi-indicator-extreme-readings/
http://marketsci.wordpress.com/2010/07/27/css-analytics%E2%80%99-dvi-indicator-revealed/
http://dvindicators.cssanalytics.com/community/?vasthtmlaction=viewtopic&t=47.0
http://quantingdutchman.wordpress.com/2010/07/28/dvi-indicator-for-amibroker/
http://davesbrain.blogs.com/mindmoneymarkets/2010/08/dvib-combo-.html
}

