% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{env}
\alias{env}
\alias{env.del}
\title{Shortcut for new.env function}
\usage{
env(..., hash = TRUE, parent = parent.frame(), size = 29L)

env.del(names, env)
}
\arguments{
\item{...}{members of new environment}
}
\value{
new environment
}
\description{
This function is a convenience shortcut for new.env function
adding ability to work with environment the same way you work with list
}
\examples{
\dontrun{ 
whitespace = ' '
env(a=10,b=3,whitespace)
env(a=10,b=3,10)
}
}
