% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ev.ratio}
\alias{ev.ratio}
\title{Rolling EV Ratio: A Trend Indicator or Performance Measurement Statistic
cumulative W% (up periods/total periods) x W/L ratio (sum of wins/sum of losses)
http://cssanalytics.wordpress.com/2010/06/02/rolling-ev-ratio-a-trend-indicator-or-performance-measurement-statistic/
http://davesbrain.blogs.com/mindmoneymarkets/2010/07/will-mean-reversion-bounce-back.html}
\usage{
ev.ratio(data, n = 252)
}
\description{
Rolling EV Ratio: A Trend Indicator or Performance Measurement Statistic
cumulative W% (up periods/total periods) x W/L ratio (sum of wins/sum of losses)
http://cssanalytics.wordpress.com/2010/06/02/rolling-ev-ratio-a-trend-indicator-or-performance-measurement-statistic/
http://davesbrain.blogs.com/mindmoneymarkets/2010/07/will-mean-reversion-bounce-back.html
}

