% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{event.at}
\alias{event.at}
\title{Helper functions, subject to change}
\usage{
event.at(x, period = "month", amount = 1, period.ends = date.ends(x,
  period), offset = 1)
}
\description{
Helper functions, subject to change
}
