% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flip.xts}
\alias{flip.xts}
\title{Reverse order of \code{\link{xts}} object}
\usage{
flip.xts(x)
}
\arguments{
\item{x}{\code{\link{xts}} object}
}
\value{
\code{\link{xts}} object
}
\description{
Reverse order of \code{\link{xts}} object
}

