% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{fred.fx.symbol}
\alias{fred.fx.symbol}
\title{Get list of FX symbols from FRED
[FRED H.10 Foreign Exchange Rates](https://research.stlouisfed.org/fred2/release?rid=17)}
\usage{
fred.fx.symbol()
}
\description{
Get list of FX symbols from FRED
[FRED H.10 Foreign Exchange Rates](https://research.stlouisfed.org/fred2/release?rid=17)
}
\examples{
\dontrun{ 
info = fred.fx.symbol()
info$fx$symbol
}
}
