% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{fxhistoricaldata.fx.symbol}
\alias{fxhistoricaldata.fx.symbol}
\title{Get list of FX symbols from FXHISTORICALDATA.COM
[FXHISTORICALDATA.COM](http://www.fxhistoricaldata.com/)}
\usage{
fxhistoricaldata.fx.symbol()
}
\description{
Get list of FX symbols from FXHISTORICALDATA.COM
[FXHISTORICALDATA.COM](http://www.fxhistoricaldata.com/)
}
\examples{
\dontrun{ 
info = fxhistoricaldata.fx.symbol()
info
}
}
