% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{getSymbols.intraday}
\alias{getSymbols.intraday}
\title{Helper function to add Intraday prices}
\usage{
getSymbols.intraday(
  Symbols = NULL,
  env = parent.frame(),
  getSymbols.fn = getSymbols,
  auto.assign = T,
  ...
)
}
\description{
Helper function to add Intraday prices
}
