% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{hist.quotes.url}
\alias{hist.quotes.url}
\title{URL for various data providers}
\usage{
\method{hist}{quotes.url}(
  ticker = "IBM",
  from = "1900-01-01",
  to = Sys.Date(),
  src = spl("yahoo,google,quotemedia")
)
}
\description{
[lookup ticker](http://www.quotemedia.com/portal/quote?qm_symbol=pot:ca)
[check sector / industry info for any company](http://www.quotemedia.com/portal/profile?qm_symbol=m)
}
\details{
hist = read.xts(hist.quotes.url('IBM', '1992-11-01', '2016-05-05', 'quotemedia'))
hist = read.xts(hist.quotes.url('HOU:CA', '1992-11-01', '2016-05-05', 'quotemedia'))
hist = read.xts(get.url(hist.quotes.url('HOD:CA', '1992-11-01', '2016-05-05', 'quotemedia')))

library(readr)
hist = read.xts(read_csv(get.url(hist.quotes.url('HOU:CA', '1992-11-01', '2016-05-05', 'quotemedia')),,na=c('','NA','N/A')))

http://web.tmxmoney.com/pricehistory.php?qm_page=90043&qm_symbol=HOD
http://www.quotemedia.com/portal/history?qm_symbol=HOD:CA
}
