% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ifna.prev}
\alias{ifna.prev}
\title{Backfill NA's with last non NA value, similar to na.locf(y,na.rm=FALSE)
http://r.789695.n4.nabble.com/Vector-replace-0-elements-without-using-a-loop-td2}
\usage{
ifna.prev(y)
}
\description{
Backfill NA's with last non NA value, similar to na.locf(y,na.rm=FALSE)
http://r.789695.n4.nabble.com/Vector-replace-0-elements-without-using-a-loop-td2
}

