% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{ls.f}
\alias{ls.f}
\alias{ls.v}
\title{List function / variables in environment}
\usage{
ls.f(env = sys.frame(-1))

ls.v(env = sys.frame(-1))
}
\description{
http://www.mail-archive.com/r-help@stat.math.ethz.ch/msg22679.html
}
