% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{lst}
\alias{lst}
\title{Shortcut for list creation function}
\usage{
lst(...)
}
\arguments{
\item{...}{members of list}
}
\value{
list
}
\description{
This function is a shortcut for list creation function
}
\examples{
\dontrun{ 
a = 1
lst(a,b=2)
}
}
