\name{make.stock.xts}
\alias{make.stock.xts}
\title{Create stock like \code{\link{xts}} object from one column time series}
\usage{
  make.stock.xts(out, column = 1)
}
\arguments{
  \item{out}{\code{\link{xts}} time series}

  \item{column}{column index to use, \strong{defaults to
  1}}
}
\value{
  stock like \code{\link{xts}} object
}
\description{
  Create stock like \code{\link{xts}} object from one
  column time series
}

