% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{map2vector}
\alias{map2vector}
\title{Flexible utility function to Map value(s) to vector}
\usage{
map2vector(expr, labels, default = 0)
}
\arguments{
\item{labels}{names of vector columns}

\item{default}{default value, \strong{defaults to 0}}

\item{value}{value(s) to use for mapping}
}
\value{
vector

Possible scenarios:
one number, applied to all assets
array, same number of entries as assets
named list, each name corresponds asset
plus have a fallback asset if nothing provided
}
\description{
Flexible utility function to Map value(s) to vector
}
\examples{
\dontrun{ 
map2vector(1, 5)
map2vector(c(1,2,3,4,5), 5)
map2vector(c(1,2,NA,4,5), 5, 100)
map2vector(list(a=1,b=4), 'a,b,c,d,e', 100)
map2vector(list(a=1,C=4), 'a,b,c,d,e', 100)
map2vector(labels= 'a,b,c,d,e', 100)
map2vector({
  a=10
  b=2
  },'a,b,c', 100)
}
}
