% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mat}
\alias{mat}
\title{convenience shortcut for matrix function}
\usage{
mat(x, col = T)
}
\arguments{
\item{x}{object}

\item{col}{flag to indicate 1 column if x is a vector; otherwise 1 row}
}
\value{
new matrix
}
\description{
convenience shortcut for matrix function
}
