% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mlag}
\alias{mlag}
\title{Lag matrix or vector}
\usage{
mlag(m, nlag = 1)
}
\arguments{
\item{nlag}{number of lags, \strong{defaults to 1}}

\item{x}{vector / matrix}
}
\value{
modified object
}
\description{
This function shifts elements in a vector or a matrix by a given lag.
For example: mlag(x,1) - use yesterday's values and
 mlag(x,-1) - use tomorrow's values
}
\examples{
\dontrun{ 
mlag(1:10)
}
}
