% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mlast}
\alias{mlast}
\title{Last elements of matrix or vector}
\usage{
mlast(m, nlast = 1)
}
\arguments{
\item{m}{vector / matrix}

\item{n}{number of elements from the end, \strong{defaults to 1}}
}
\value{
last n elements
}
\description{
Last elements of matrix or vector
}
\examples{
\dontrun{ 
mlast(1:10)
}
}
