% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ntop.keep}
\alias{ntop.keep}
\title{Select top N for each period, and keep them till they drop below keepn rank
http://www.etfscreen.com/sectorstrategy.php}
\usage{
ntop.keep(data, topn = 1, keepn = 1, dirMaxMin = TRUE)
}
\description{
Select top N for each period, and keep them till they drop below keepn rank
http://www.etfscreen.com/sectorstrategy.php
}

