% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ntop}
\alias{ntop}
\title{Sample rotation Strategies
Select top N for each period
http://www.etfscreen.com/sectorstrategy.php}
\usage{
ntop(data, topn = 1, dirMaxMin = TRUE)
}
\description{
Sample rotation Strategies
Select top N for each period
http://www.etfscreen.com/sectorstrategy.php
}

