% Generated by roxygen2 (4.0.1): do not edit by hand
\name{percent.rank}
\alias{percent.rank}
\title{Percentile Rank over given window, works for both single column and matrix
If data is matrix, lookup value in the first column across all columns}
\usage{
percent.rank(data, n = 252)
}
\description{
Percentile Rank over given window, works for both single column and matrix
If data is matrix, lookup value in the first column across all columns
}

