% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.r
\name{plot.txt}
\alias{plot.txt}
\title{Create Ascii(txt) plot}
\usage{
\method{plot}{txt}(y, x = NULL, pch = 1:ncol(y),
  width = round(options()$width * 0.8))
}
\arguments{
\item{y}{matrix or vector of y values}

\item{x}{optional vector of x values, \strong{defaults to 1:nr, where nr is the number of rows in y}}

\item{pch}{optional vector of markers for time series, \strong{defaults to 1:nc, where nc is the number of columns in y}}
}
\value{
nothing
}
\description{
Based on the following idea: [ASCII Plotting Functions for R](http://stackoverflow.com/questions/14736556/ascii-plotting-functions-for-r)
}
\details{
Plot multiple series by stacking data and using corresponding labels (pch)
also works with NA's so we can potentially plot many partial series
}
\examples{
\dontrun{ 
txtplot(1:10, c(1,2,3,4,5,NA,7,8,NA,10))

x = 1:10;y = 1:10;y1 = 30:21
txtplot(rbind(x,x), rbind(y,y1), pch=c('+','*'))

plot.txt(cbind(y,y1),x)
plot.txt(cbind(y,y1),x,pch=c('+','*'))
}
}
