% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.r
\name{plota.format}
\alias{plota.format}
\title{Format numbers using 1000 separator}
\usage{
plota.format(temp, nround = 2, sprefix = "", eprefix = "")
}
\arguments{
\item{temp}{numbers}

\item{nround}{number of rounding digits, \strong{defaults to '2'}}

\item{sprefix}{start prefix string, \strong{defaults to ''}}

\item{eprefix}{end postfix string, \strong{defaults to ''}}
}
\value{
formatted numbers
}
\description{
Format numbers using 1000 separator
}
