\name{plota.lines}
\alias{plota.lines}
\title{Add lines to time series plot}
\usage{
  plota.lines(y, type = "l", col = par("col"), ...)
}
\arguments{
  \item{y}{\code{\link{xts}} object}

  \item{type}{line type, \strong{defaults to 'l'}, for more
  info see \code{\link{lines}}}

  \item{col}{color, \strong{defaults to par('col')}}

  \item{...}{additional parameters to the
  \code{\link{lines}}}
}
\value{
  nothing
}
\description{
  Add lines to time series plot
}

