% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plota.r
\name{plota.text}
\alias{plota.text}
\title{Add text to time series plot}
\usage{
plota.text(y, ...)
}
\arguments{
\item{y}{\code{\link{xts}} object}

\item{...}{additional parameters to the \code{\link{lines}}}
}
\value{
nothing
}
\description{
Add text to time series plot
}
