\name{plota.x.highlight}
\alias{plota.x.highlight}
\title{Highlight vertical segments}
\usage{
  plota.x.highlight(y, highlight,
    col = plota.control$col.x.highlight)
}
\arguments{
  \item{y}{\code{\link{xts}} object}

  \item{highlight}{segments to highlight along X axis}

  \item{col}{highlight color, \strong{defaults to
  plota.control$col.x.highlight}}
}
\value{
  nothing
}
\description{
  Highlight vertical segments
}

