\name{repmat}
\alias{repmat}
\title{Replicate and tile a given vector}
\usage{
  repmat(v, n, m)
}
\arguments{
  \item{v}{vector}

  \item{n}{number of copies along rows}

  \item{m}{number of copies along columns}
}
\value{
  new matrix
}
\description{
  Replicate and tile a given vector
}
\examples{
\dontrun{
repmat(1:10,1,2)
}
}
\references{
  \url{http://www.mathworks.com/help/techdoc/ref/repmat.html}
}

