% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{round.lot.basic.base}
\alias{round.lot.basic.base}
\title{Round Lot Basic Base
round lot to the nearest lot.size}
\usage{
\method{round}{lot.basic.base}(weight, price, capital, lot.size)
}
\description{
Round Lot Basic Base
round lot to the nearest lot.size
}
