% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{round.lot.basic}
\alias{round.lot.basic}
\title{Round Lot Basic
round lot to the nearest lot.size and next try to reallocate remaining cash}
\usage{
\method{round}{lot.basic}(weight, price, capital, lot.size)
}
\description{
Round Lot Basic
round lot to the nearest lot.size and next try to reallocate remaining cash
}
