% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt.share.r
\name{round.lot}
\alias{round.lot}
\title{Round Lot
a helper function for round.lot.basic}
\usage{
\method{round}{lot}(weight, price, capital, lot.size)
}
\description{
Round Lot
a helper function for round.lot.basic
}
