\name{spl}
\alias{spl}
\title{Split string into tokens using delim}
\usage{
  spl(s, delim = ",")
}
\arguments{
  \item{s}{input string}

  \item{delim}{delimiter, \strong{defaults to ","}}
}
\value{
  array of tokens
}
\description{
  This function will split given string into tokens using
  delim
}
\examples{
\dontrun{
spl('a,b,c')
}
}

