% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{string.buffer}
\alias{string.buffer}
\alias{add}
\alias{add.StringBuffer}
\alias{string}
\alias{string.StringBuffer}
\alias{close}
\alias{close.StringBuffer}
\title{String Buffer class - fast storage for strings}
\usage{
\method{string}{buffer}()

add(x, ..., sep, end.sep)

\method{add}{StringBuffer}(x, ..., sep = ",", end.sep = "\\n")

string(x)

\method{string}{StringBuffer}(x)

close(x)

\method{close}{StringBuffer}(x)
}
\description{
String Buffer class - fast storage for strings
}
\examples{
\dontrun{ 
sb = string.buffer()
add(sb, 'asbcd')
add(sb, '234543')
string(sb)
close(sb)
sb=NULL
}
}
