% Generated by roxygen2 (4.0.1): do not edit by hand
\name{super.smoother.filter}
\alias{super.smoother.filter}
\title{SuperSmoother filter 2013 John F. Ehlers
http://www.stockspotter.com/files/PredictiveIndicators.pdf}
\usage{
super.smoother.filter(x)
}
\description{
SuperSmoother filter 2013 John F. Ehlers
http://www.stockspotter.com/files/PredictiveIndicators.pdf
}

