\name{tableColor}
\alias{tableColor}
\title{Add color highlighting to the \code{\link{reactiveTable}} function}
\usage{
  tableColor(data, header.col = "LightGray",
    row.col = "yellow", negative.col = "red", ...)
}
\arguments{
  \item{data}{table data}

  \item{header.col}{background color for the header row}

  \item{row.col}{background color for alternative rows}

  \item{negative.col}{text color for negative values}

  \item{...}{additional parameters to the
  \code{\link{reactiveTable}} fucntion}
}
\value{
  HTML table
}
\description{
  Add color highlighting to the \code{\link{reactiveTable}}
  function
}

