% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{test.equality}
\alias{test.equality}
\title{Shortcut for all.equal function}
\usage{
test.equality(..., eps = 1e-10, na.rm = T, type = c("first", "all"))
}
\arguments{
\item{...}{variables to compare}
}
\value{
matrix with results of comparison

all.equal(r.cor, c.cor)
all(abs(r.cor - c.cor) < 1e-10, na.rm=T)
}
\description{
This function is a convenience shortcut for all.equal function
adding ability to work with multiple variables
}
