% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{tic}
\alias{tic}
\alias{toc}
\title{Begin Timing}
\usage{
tic(identifier)

toc(identifier)
}
\arguments{
\item{identifier}{name for this timing session}
}
\value{
nothing

elapsed time
}
\description{
Begin Timing

End Timing and report elapsed time
}
\examples{
\dontrun{ 
tic(1)
}
\dontrun{ 
toc(1)
}
}
