% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{trim}
\alias{trim}
\title{Remove any leading and trailing spaces}
\usage{
trim(s)
}
\arguments{
\item{s}{string}
}
\value{
resulting string
}
\description{
This function will remove any leading and trailing spaces
}
\examples{
\dontrun{ 
trim('  a b c  ')
}
}
