% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ulcer.index}
\alias{ulcer.index}
\title{Ulcer Index: alternative to Standard Devation
http://en.wikipedia.org/wiki/Ulcer_index
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:ulcer_index}
\usage{
ulcer.index(x, n = 14)
}
\description{
Ulcer Index: alternative to Standard Devation
http://en.wikipedia.org/wiki/Ulcer_index
http://stockcharts.com/school/doku.php?id=chart_school:technical_indicators:ulcer_index
}

