% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.r
\name{visualize.system.parameter.optimization}
\alias{visualize.system.parameter.optimization}
\title{Visualize System Parameter Optimization
based on [Visualizing Data](http://sanzprophet.blogspot.tw/2013/01/visualizing-data.html)
The article was using [XDat](http://www.xdat.org/index.php?ref=download) app}
\usage{
visualize.system.parameter.optimization(result)
}
\value{
nothing
}
\description{
Visualize System Parameter Optimization
based on [Visualizing Data](http://sanzprophet.blogspot.tw/2013/01/visualizing-data.html)
The article was using [XDat](http://www.xdat.org/index.php?ref=download) app
}
\examples{
\dontrun{ 
result = cbind(rnorm(20,0,1),rnorm(20,10,1),rnorm(20,0,100),rnorm(20,20,21))
 colnames(result) = spl('CAGR,A,B,C')
visualize.system.parameter.optimization(result)
}
}
