% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{blocksum}
\alias{blocksum}
\title{Compute the block-wise sum of a vector.}
\usage{
blocksum(r, c)
}
\arguments{
\item{r}{An integer vector}

\item{c}{The number of observations in each block}
}
\value{
The function returns the block sum of the vector.
}
\description{
Compute the block-wise sum of a vector.
}
