% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.proxy.r
\name{compute.cor}
\alias{compute.cor}
\title{Compute correlations}
\usage{
compute.cor(data, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{data}{matrix with data}

\item{method}{method used to compute correlations, please see \code{\link{cor}} for more details}
}
\value{
correlation matrix
}
\description{
Compute correlations
}
