\name{get.extension}
\alias{get.extension}
\alias{get.filename}
\alias{get.full.filename}
\title{File name Functions}
\usage{
  get.extension(x)

  get.full.filename(x)

  get.filename(x)
}
\arguments{
  \item{x}{file name}
}
\value{
  part of the file name
}
\description{
  File name Functions
}
\examples{
\dontrun{
get.extension('test.csv')
}
}

