% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.r
\name{iline}
\alias{iline}
\title{Add Vertical, Horizontal, Cross line to existing plot at location of mouse click
based on getGraphicsEvent {grDevices}}
\usage{
iline(
  type = c("v", "h", "cross"),
  col = "red",
  remove.col = "white",
  stop.key = "q"
)
}
\value{
nothing
}
\description{
Please note there is also an interactive plot version in
library(playwith)
playwith(plot(1:10))
}
\examples{
\dontrun{ 
plot(1:10)
iline()
}
}
