\name{plota.format}
\alias{plota.format}
\title{Format numbers using 1000 separator}
\usage{
  plota.format(temp, nround = 2, sprefix = "",
    eprefix = "")
}
\arguments{
  \item{temp}{numbers}

  \item{nround}{number of rounding digits, \strong{defaults
  to '2'}}

  \item{sprefix}{start prefix string, \strong{defaults to
  ''}}

  \item{eprefix}{end postfix string, \strong{defaults to
  ''}}
}
\value{
  formated numbers
}
\description{
  Format numbers using 1000 separator
}

