\name{xts2ts}
\alias{xts2ts}
\title{Convert \code{\link{xts}} object to \code{\link{ts}} object}
\usage{
  xts2ts(x)
}
\arguments{
  \item{x}{\code{\link{ts}} object}
}
\value{
  \code{\link{xts}} object

  use as.xts(ts) to convert back
}
\description{
  Convert \code{\link{xts}} object to \code{\link{ts}}
  object
}

